app/requirements.txt
sqlalchemy==1.4.43
psycopg2-binary==2.9.3
Flask==2.0.2
Werkzeug==2.2.2